Fix dependency graph
We now treat each package in a workspace as effectively a "dependency file".

Only one instance of each dependency is included in the list of
dependencies, as otherwise this can explode very quickly.

Because of this, we can no longer provide a single dependency path for
each dependency, so this field has been dropped.